+++ numpy-sys/src/arrayiter.rs
@@ -3,7 +3,7 @@
 
 use pyffi::*;
 use super::types::*;
-use super::pyarray::*;
+use super::array::*;
 
 pub type npy_bool = ::std::os::raw::c_uchar;
 pub type npy_iter_get_dataptr_t =
