+++ numpy-sys/src/lib.rs
@@ -3,6 +3,6 @@ extern crate libc;
 extern crate python3_sys as pyffi;
 
 pub mod types;
-pub mod pyarray;
-pub mod pyarrayiter;
-pub mod pyufunc;
+pub mod array;
+pub mod arrayiter;
+pub mod ufunc;
